requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.53.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-base-url==2.1.0
pytest-bdd==8.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
python-slugify==8.0.4
requests==2.32.4
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==2.5.0
testrail-api>=1.13.4
python-dotenv==1.1.1
junitparser==4.0.2


